Fix some complainted choices made in my older commits. Note: probably you should do 'make clean' before using 'make' to rebuild it again. Maybe it's me but 'make' reports various errors when linking.

+++ cli/cppcheckexecutor.h
@@ -81,7 +81,7 @@ public:
      * @param sizedone The sum of sizes of the files checked.
      * @param sizetotal The total sizes of the files.
      */
-    static void reportStatus(unsigned int fileindex, unsigned int filecount, long sizedone, long sizetotal);
+    static void reportStatus(size_t fileindex, size_t filecount, long sizedone, long sizetotal);
 
 protected:
